requirements.txt
django
Mastodon.py
python-dotenv
openai
